feat: 몇분전 시간 계산 함수 util로 분리
