+++ FE/src/components/issues/IssueList.jsx
@@ -3,6 +3,7 @@ import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import { IconMilestone } from '../../assets/icons/IconMilestone';
+import { calculatePastTime } from '../../utils/dateUtils';
 
 export default function IssueList({ isSingleChecked, setCheckedItems, listData }) {
     const { id, title, createDate, milestoneName, authorId, assignees, labels } = listData;
@@ -14,24 +15,12 @@ export default function IssueList({ isSingleChecked, setCheckedItems, listData }
         else setCheckedItems((prev) => [...prev, id]);
     };
 
-    const calculatePastTime = () => {
-        const registerDate = new Date(createDate);
-        const diffDate = Math.abs(new Date() - registerDate);
-        const diffMiniutes = Math.floor(diffDate / (1000 * 60));
-        const diffHours = Math.floor(diffDate / (1000 * 60 * 60));
-        const diffDays = Math.floor(diffDate / (1000 * 60 * 60 * 24));
-
-        if (diffDays >= 1) return `${diffDays}일 전`;
-        if (diffHours >= 1) return `${diffHours}시간 전`;
-        return diffMiniutes >= 1 ? `${diffMiniutes}분 전` : `방금 전`;
-    };
-
     useEffect(() => {
-        setPastTime(calculatePastTime());
+        setPastTime(calculatePastTimelatePastTime(createDate));
 
         const intervalPerTime = () => {
             setInterval(() => {
-                setPastTime(calculatePastTime());
+                setPastTime(calculatePastTime(createDate));
             }, 1000 * 60);
         };
